Move feed_purldb code to purldb module
